Modified AudioView to check freq bin sz

+++ PhaseI/audioView/src/AudioView.hpp
@@ -112,5 +112,5 @@ class AudioView{
     void resetSpectrograph();
 
     /* Method to reformat and access spectral data */
-    std::vector<std::vector<float>> getSpectralData();
+    std::vector<std::vector<int>> getSpectralData();
 };
